add fastest level time and completed tracking in storage service

+++ lib/features/game/data/datasource/game_local_data_source.dart
@@ -5,6 +5,6 @@ abstract class GameLocalDataSource {
   Future<void> setFastestLevelTime(double time);
   Future<void> setFastestLevelCompleted(int level);
 
-  Future<int> getFastestLevelTime();
+  Future<double> getFastestLevelTime();
   Future<int> getFastestLevelCompleted();
 }
